refactor(action): reorganize singleplayer session related actions
